Extract sample hold helpers
